Removed HttpOptions
This fixes #5582
