+++ Duplicati/Library/Modules/Builtin/Strings.cs
@@ -35,28 +35,12 @@ namespace Duplicati.Library.Modules.Builtin.Strings
         public static string ForcepassphrasefromstdinLong { get { return LC.L(@"By default, the passphrase is attempted read from the TTY device directly, increasing the security by not copying the passphrase into a stream. In some setups, such as when running detached from a console, this does not work. Set this flag to prevent trying a TTY read and only read the passphrase from STDIN."); } }
         public static string ForcepassphrasefromstdinShort { get { return LC.L(@"Read passphrase from STDIN"); } }
     }
-    internal static class HttpOptions
+    internal static class CommonOptions
     {
-        public static string Description { get { return LC.L(@"This module exposes a number of properties that can be used to change the way http requests are issued"); } }
-        public static string DisplayName { get { return LC.L(@"Configure http requests"); } }
-        public static string DescriptionAcceptAnyCertificateLong { get { return LC.L(@"Use this option to accept any server certificate, regardless of what errors it may have. Please use --{0} instead, whenever possible.", "accept-specified-ssl-hash"); } }
-        public static string DescriptionAcceptAnyCertificateShort { get { return LC.L(@"Accept any server certificate"); } }
-        public static string DescriptionAcceptHashLong2 { get { return LC.L(@"If your server certificate is reported as invalid (e.g. with self-signed certificates), you can supply the certificate hash (SHA1) to approve it anyway. The hash value must be entered in hex format without spaces or colons. You can enter multiple hashes separated by commas."); } }
-        public static string DescriptionAcceptHashShort { get { return LC.L(@"Optionally accept a known SSL certificate"); } }
-        public static string DisableExpect100Long { get { return LC.L(@"The default HTTP request has the header ""Expect: 100-Continue"" attached, which allows some optimizations when authenticating, but also breaks some web servers, causing them to report ""417 - Expectation failed""."); } }
-        public static string DisableExpect100Short { get { return LC.L(@"Disable the expect header"); } }
-        public static string DisableNagleLong { get { return LC.L(@"By default the http requests use the RFC 896 nagling algorithm to support transfer of small packages more efficiently."); } }
-        public static string DisableNagleShort { get { return LC.L(@"Disable nagling"); } }
+        public static string Description { get { return LC.L(@"This module exposes a number of common properties that can be used to change the way backends are configured"); } }
+        public static string DisplayName { get { return LC.L(@"Configure backend options"); } }
         public static string OauthurlLong { get { return LC.L(@"Duplicati uses an external server to support the OAuth authentication flow. If you have set up your own Duplicati OAuth server, you can supply the refresh URL."); } }
         public static string OauthurlShort { get { return LC.L(@"Alternate OAuth URL"); } }
-        public static string SslversionsLong { get { return LC.L(@"This option changes the default SSL versions allowed. This is an advanced option and should only be used if you want to enhance security or work around an issue with a particular SSL protocol."); } }
-        public static string SslversionsShort { get { return LC.L(@"Set allowed SSL versions"); } }
-        public static string OperationtimeoutLong { get { return LC.L(@"This option changes the default timeout for any HTTP request, the time covers the entire operation from initial packet to shutdown."); } }
-        public static string OperationtimeoutShort { get { return LC.L(@"Set the default operation timeout"); } }
-        public static string ReadwritetimeoutLong { get { return LC.L(@"This option changes the default read-write timeout. Read-write timeouts are used to detect a stalled requests, and this option configures the maximum time between activity on a connection."); } }
-        public static string ReadwritetimeoutShort { get { return LC.L(@"Set readwrite"); } }
-        public static string BufferrequestsLong { get { return LC.L(@"This option sets the HTTP buffering. Setting this to ""{0}"" can cause memory leaks, but can also improve performance in some cases.", "true"); } }
-        public static string BufferrequestsShort { get { return LC.L(@"Set HTTP buffering"); } }
     }
     internal static class HyperVOptions
     {
